multi_agent/requirements.txt

crewai==1.7.2
crewai-tools==1.7.2
py-toon-format==0.1.0
python-dotenv==1.1.1